+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 1.0.1-beta2
 * `PhpDocMakerCommand` now correctly handles PHP errors too (e.g. notice);
 * added `AbstractEntity::getTagsByName()` method. Invalid tags will throw an exception;
+* added `ClassEntity::getParentClass()` method;
 * added `ConstantEntity::getValueAsString()` method, fixed bug for class costants
     with an array as value;
 * fixed little bug for `ClassEntity::getConstants()` method;